[MISC] move test adaptor_base_test
