+++ test/unit/core/range/CMakeLists.txt
@@ -1,1 +1,3 @@
 seqan3_test(range_iterator_test.cpp)
+
+add_subdirectories()
