+++ test/unit/range/views/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_subdirectories()
 
-seqan3_test(adaptor_base_test.cpp)
 seqan3_test(as_const_test.cpp)
 seqan3_test(convert_test.cpp)
 seqan3_test(deep_test.cpp)
